Разбиение на модули + правка пути entry в конфиге вебпака
