+++ src/components/footer-stats.js
@@ -1,0 +1,9 @@
+const createFooterStatsTemplate = () => {
+  return (
+    `<section class="footer__statistics">
+      <p>130 291 movies inside</p>
+    </section>`
+  );
+};
+
+export {createFooterStatsTemplate};
