+++ src/components/more-btn.js
@@ -1,0 +1,7 @@
+const createMoreBtn = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createMoreBtn};
